Moved the ScrollTop to top event to the Event class and updated the ProductList and ProductDetail View Models to use the new event.
